Merge -r1.67 of ufs/inode.h: make the inode field of a filehandle 64-bit instead of truncating to 32 bits. Note that if you're serving nfs off lfs (but I don't think you are as I think there are known fatal problems doing so) you'll need to reboot your clients after this change.
I've used a 64-bit value explicitly instead of ino_t (as in the ufs
structure) because this is a structure whose size ought to be well
defined. I remember some discussion of this when the ufs change was
committed, but not the conclusion (if any) -- if anyone hates this it
can be changed to ino_t easily enough.