Add new tests dbregs[2345] in MD arch/amd64/ t_ptrace_wait{,3,4,6,id,pid}
dbregs2:
    Verify that setting DR0 is preserved across ptrace(2) calls

dbregs3:
    Verify that setting DR1 is preserved across ptrace(2) calls

dbregs4:
    Verify that setting DR2 is preserved across ptrace(2) calls

dbregs5:
    Verify that setting DR3 is preserved across ptrace(2) calls

These tests are deliberately fine-grained as they are expected to penetrate
precisely each functional aspect of CPU Debug Registers on amd64 one after
another.

These tests (and MI ones) might be generated or merged with helper
functions, however in order to copy-and-paste them out of a test-suite and
quickly port to other platform (in order to compare results) it's useful to
keep them as stand-alone as they are.

Code from these tests might be shared with other ports in future, for the
same reason keep them currently as they are.

Sponsored by <The NetBSD Foundation>